1.10.5.4

* Target fps now allows floats instead of integers for more fluent animations, resulting in smoother scrolling etc.
* Scrolling ending jagging removed on density < 1.5 devices.
* SkiaLayout Wrap fix for auto-sized children

+++ src/Maui/DrawnUi/Platforms/MacCatalyst/Super.Mac.cs
@@ -109,12 +109,12 @@ namespace DrawnUi.Draw
 
         #endregion
 
-        public static int GetDisplayRefreshRate(int fallback)
+        public static float GetDisplayRefreshRate(float fallback)
         {
             return (int)UIScreen.MainScreen.MaximumFramesPerSecond;
         }
 
-        public static int RefreshRate { get; protected set; }
+        public static float RefreshRate { get; protected set; }
 
 
         public static void Init()
