+++ src/Maui/DrawnUi/Platforms/iOS/Super.iOS.cs
@@ -15,12 +15,12 @@ namespace DrawnUi.Draw
     public partial class Super
     {
 
-        public static int GetDisplayRefreshRate(int fallback)
+        public static float GetDisplayRefreshRate(float fallback)
         {
             return (int)UIScreen.MainScreen.MaximumFramesPerSecond;
         }
 
-        public static int RefreshRate { get; protected set; }
+        public static float RefreshRate { get; protected set; }
 
         public static void Init()
         {
